Handle relative module source paths
This commit updates the plugin to support relative module source paths.
A relative path begins with "." and denotes a direct path to a
"package.json" metadata file. Such paths are resolved so that an
absolute unique identifier can be used to refer to the bundle loaded
from such a module.